pyproject.toml
[project]
name = "proxy-mcp-server"
version = "0.1.0"
description = "MCP Server acting as a proxy for SDR Backend"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "starlette>=0.32.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]


